requirements.txt
Django==4.2.16
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
Pillow==10.4.0
gunicorn==22.0.0
